chore: 파일 구조 정리
