--- src/main/java/com/programmers/cafekiosk/dto/CreateOrderRequest.java
@@ -1,6 +1,0 @@
-package com.programmers.cafekiosk.dto;
-
-import java.util.List;
-
-public record CreateOrderRequest(List<CreateOrderItemRequest> orderitems) {
-}
