--- src/main/java/com/programmers/cafekiosk/dto/GetItemsRequest.java
@@ -1,18 +1,0 @@
-package com.programmers.cafekiosk.dto;
-
-import com.programmers.cafekiosk.entity.ItemStatus;
-import com.programmers.cafekiosk.entity.ItemType;
-
-import java.util.Set;
-
-public record GetItemsRequest(
-
-        @ValidItemType
-        ItemType type,
-
-        @ValidItemStatus
-        ItemStatus status,
-
-        Set<Long> ids
-) {
-}
