+++ src/main/java/com/programmers/cafekiosk/dto/item/request/UpdateItemRequest.java
@@ -1,7 +1,9 @@
-package com.programmers.cafekiosk.dto;
+package com.programmers.cafekiosk.dto.item.request;
 
-import com.programmers.cafekiosk.entity.ItemStatus;
-import com.programmers.cafekiosk.entity.ItemType;
+import com.programmers.cafekiosk.dto.item.validator.ValidItemStatus;
+import com.programmers.cafekiosk.dto.item.validator.ValidItemType;
+import com.programmers.cafekiosk.entity.item.ItemStatus;
+import com.programmers.cafekiosk.entity.item.ItemType;
 import jakarta.validation.constraints.NotNull;
 
 public record UpdateItemRequest(
