+++ src/main/java/com/programmers/cafekiosk/dto/item/validator/ValidItemStatus.java
@@ -1,5 +1,6 @@
-package com.programmers.cafekiosk.dto;
+package com.programmers.cafekiosk.dto.item.validator;
 
+import com.programmers.cafekiosk.dto.item.validator.ItemStatusValidator;
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
 
@@ -7,13 +8,12 @@ import java.lang.annotation.*;
 
 @Target({ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = ItemTypeValidator.class)
+@Constraint(validatedBy = ItemStatusValidator.class)
 @Documented
-public @interface ValidItemType {
-    String message() default "Invalid ItemType";
+public @interface ValidItemStatus {
+    String message() default "Invalid ItemStatus";
 
     Class<?>[] groups() default {};
 
     Class<? extends Payload>[] payload() default {};
 }
-
