+++ src/main/java/com/programmers/cafekiosk/dto/item/validator/ValidItemType.java
@@ -1,5 +1,6 @@
-package com.programmers.cafekiosk.dto;
+package com.programmers.cafekiosk.dto.item.validator;
 
+import com.programmers.cafekiosk.dto.item.validator.ItemTypeValidator;
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
 
@@ -7,12 +8,13 @@ import java.lang.annotation.*;
 
 @Target({ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = ItemStatusValidator.class)
+@Constraint(validatedBy = ItemTypeValidator.class)
 @Documented
-public @interface ValidItemStatus {
-    String message() default "Invalid ItemStatus";
+public @interface ValidItemType {
+    String message() default "Invalid ItemType";
 
     Class<?>[] groups() default {};
 
     Class<? extends Payload>[] payload() default {};
 }
+
