+++ src/main/java/com/programmers/cafekiosk/dto/order/CreateOrderRequest.java
@@ -1,0 +1,8 @@
+package com.programmers.cafekiosk.dto.order;
+
+import com.programmers.cafekiosk.dto.orderItem.CreateOrderItemRequest;
+
+import java.util.List;
+
+public record CreateOrderRequest(List<CreateOrderItemRequest> orderitems) {
+}
