+++ src/main/java/com/programmers/cafekiosk/dto/order/OrderResponse.java
@@ -1,4 +1,6 @@
-package com.programmers.cafekiosk.dto;
+package com.programmers.cafekiosk.dto.order;
+
+import com.programmers.cafekiosk.dto.orderItem.OrderItemResponse;
 
 import java.time.LocalDateTime;
 import java.util.List;
