--- src/main/java/com/programmers/cafekiosk/entity/ItemType.java
@@ -1,5 +1,0 @@
-package com.programmers.cafekiosk.entity;
-
-public enum ItemType {
-    BEVERAGE, FOOD
-}
