+++ src/main/java/com/programmers/cafekiosk/entity/item/ItemStatus.java
@@ -1,4 +1,4 @@
-package com.programmers.cafekiosk.entity;
+package com.programmers.cafekiosk.entity.item;
 
 public enum ItemStatus {
     AVAILABLE, SOLD_OUT, DISCONTINUED
