+++ src/main/java/com/programmers/cafekiosk/entity/item/ItemType.java
@@ -1,0 +1,5 @@
+package com.programmers.cafekiosk.entity.item;
+
+public enum ItemType {
+    BEVERAGE, FOOD
+}
