+++ src/main/java/com/programmers/cafekiosk/repository/item/ItemRepository.java
@@ -1,6 +1,6 @@
-package com.programmers.cafekiosk.repository;
+package com.programmers.cafekiosk.repository.item;
 
-import com.programmers.cafekiosk.entity.Item;
+import com.programmers.cafekiosk.entity.item.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ItemRepository extends JpaRepository<Item, Long> {
